+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -17,6 +17,10 @@
 - Better handling of Unicode Supplementary Planes
 - Incorrect processor group could be assigned to programs stated in `.bridge.json`
 
+#### Changed
+- Performance improvements
+- Replaced configuration pop-ups with CodeLens
+
 ## [1.8.0](https://github.com/eclipse-che4z/che-che4z-lsp-for-hlasm/compare/1.7.0...1.8.0) (2023-05-24)
 
 #### Added